test(vue): unit test for script setup block
